Separate main Brobot class into BrobotGui and Brobot Cli
