+++ src/main/java/Launcher.java
@@ -5,6 +5,6 @@ import javafx.application.Application;
  */
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(Brobot.class, args);
+        Application.launch(BrobotGui.class, args);
     }
 }
